Remove RISC-V QEMU Virt platform files and update references
- Deleted axconfig.toml, build.rs, linker.lds.S, boot.rs, console.rs, init.rs, irq.rs, lib.rs, mem.rs, power.rs, and time.rs files for the riscv64-qemu-virt platform.
- Updated the platform package name from "axplat-riscv64-qemu-virt-hv" to "ax-plat-riscv64-qemu-virt" in build scripts and tests.
- Removed all associated code and configurations related to the RISC-V QEMU Virt platform.
