+++ Cargo.toml
@@ -204,7 +204,6 @@ members = [
     "platform/axplat-dyn",
     "platform/somehal",
     "platform/riscv64-visionfive2",
-    "platform/riscv64-qemu-virt",
     "platform/x86-qemu-q35",
     "scripts/axbuild",
     "test-suit/arceos/rust/backtrace-raw-basic",
@@ -358,7 +357,6 @@ axhvc = { version = "0.4.7", path = "components/axhvc" }
 axklib = { version = "0.5.8", path = "components/axklib" }
 axplat-dyn = { version = "0.6.1", path = "platform/axplat-dyn", default-features = false }
 axplat-riscv64-visionfive2 = { version = "0.1.3", path = "platform/riscv64-visionfive2" }
-axplat-riscv64-qemu-virt-hv = { version = "0.5.7", path = "platform/riscv64-qemu-virt" }
 axplat-x86-qemu-q35 = { version = "0.4.7", path = "platform/x86-qemu-q35" }
 axpoll = { version = "0.3.9", path = "components/axpoll" }
 axvcpu = { version = "0.5.8", path = "components/axvcpu" }
