+++ components/axplat_crates/platforms/axplat-riscv64-qemu-virt/Cargo.toml
@@ -11,6 +11,7 @@ license = "Apache-2.0"
 
 [features]
 fp-simd = ["ax-cpu/fp-simd"]
+hypervisor = ["irq", "smp"]
 irq = ["ax-plat/irq", "dep:ax-riscv-plic"]
 rtc = ["dep:riscv_goldfish"]
 smp = ["ax-plat/smp"]
