+++ docs/docs/development/arceos.md
@@ -387,9 +387,8 @@ components/axplat_crates/platforms/axplat-aarch64-qemu-virt/
 
 | 目录 | 内容 |
 |------|------|
-| `components/axplat_crates/platforms/` | 8 个工作区内平台 crate |
+| `components/axplat_crates/platforms/` | 工作区内 `ax-plat-*` 平台 crate |
 | `platform/axplat-dyn/` | 动态平台加载（设备树驱动） |
-| `platform/riscv64-qemu-virt/` | RISC-V QEMU virt 独立平台 |
 | `platform/x86-qemu-q35/` | x86_64 QEMU Q35 独立平台 |
 
 已有平台：
