+++ os/axvisor/src/hal/arch/riscv64/api.rs
@@ -1,8 +1,5 @@
-struct VmInterruptIfImpl;
-
-#[ax_crate_interface::impl_interface]
-impl axplat_riscv64_qemu_virt_hv::irq::InjectIrqIf for VmInterruptIfImpl {
-    fn inject_virtual_interrupt(irq: usize) {
-        crate::hal::arch::inject_interrupt(irq);
-    }
+pub(super) fn init_platform_irq_injector() {
+    ax_plat_riscv64_qemu_virt::irq::register_virtual_irq_injector(
+        crate::hal::arch::inject_interrupt,
+    );
 }
