+++ os/axvisor/src/hal/arch/riscv64/mod.rs
@@ -2,11 +2,12 @@ mod api;
 pub mod cache;
 
 use crate::vmm::vm_list::get_vm_by_id;
+use ax_plat_riscv64_qemu_virt::config::devices::PLIC_PADDR;
 use axaddrspace::{GuestPhysAddr, device::AccessWidth};
-use axplat_riscv64_qemu_virt_hv::config::devices::PLIC_PADDR;
 use axvisor_api::vmm::current_vm_id;
 
 pub fn hardware_check() {
+    api::init_platform_irq_injector();
     // TODO: implement hardware checks for RISC-V64
     // check page table level like aarch64
 }
