--- platform/riscv64-qemu-virt/CHANGELOG.md
@@ -1,36 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## [Unreleased]
-
-## [0.5.7](https://github.com/rcore-os/tgoskits/compare/axplat-riscv64-qemu-virt-hv-v0.5.6...axplat-riscv64-qemu-virt-hv-v0.5.7) - 2026-05-19
-
-### Other
-
-- updated the following local packages: ax-riscv-plic, axvisor_api, ax-cpu
-
-## [0.5.6](https://github.com/rcore-os/tgoskits/compare/axplat-riscv64-qemu-virt-hv-v0.5.5...axplat-riscv64-qemu-virt-hv-v0.5.6) - 2026-05-15
-
-### Added
-
-- *(drivers)* migrate Sparreal driver crates ([#540](https://github.com/rcore-os/tgoskits/pull/540))
-- *(irq)* pass IRQ/event number to registered handlers
-- *(console)* add interrupt-driven console input ([#343](https://github.com/rcore-os/tgoskits/pull/343))
-
-### Fixed
-
-- *(console)* keep UART writes raw ([#402](https://github.com/rcore-os/tgoskits/pull/402))
-
-### Other
-
-- *(platform)* inherit workspace metadata
-
-## [0.5.5](https://github.com/rcore-os/tgoskits/compare/axplat-riscv64-qemu-virt-hv-v0.5.4...axplat-riscv64-qemu-virt-hv-v0.5.5) - 2026-04-27
-
-### Other
-
-- updated the following local packages: ax-cpu
