--- platform/riscv64-qemu-virt/Cargo.toml
@@ -1,38 +1,0 @@
-[package]
-edition.workspace = true
-name = "axplat-riscv64-qemu-virt-hv"
-version = "0.5.7"
-description = "Hardware platform implementation for the RISC-V64 QEMU virt machine used by AxVisor"
-repository = "https://github.com/rcore-os/tgoskits"
-keywords = ["axvisor", "riscv64", "qemu", "virt", "hypervisor"]
-categories = ["embedded", "hardware-support", "no-std", "os"]
-license = "Apache-2.0"
-
-[features]
-default = [
-  "irq",
-  "smp",
-]
-fp-simd = ["ax-cpu/fp-simd"]
-irq = ["ax-plat/irq"]
-rtc = ["riscv_goldfish"]
-smp = ["ax-plat/smp"]
-
-[target.'cfg(target_arch = "riscv64")'.dependencies]
-ax-kspin = { workspace = true }
-ax-lazyinit = { workspace = true }
-log = "0.4"
-riscv = "0.14.0"
-riscv_goldfish = { version = "0.1", optional = true }
-ax-riscv-plic = { workspace = true }
-sbi-rt = { version = "0.0.3", features = ["legacy"] }
-some-serial.workspace = true
-
-ax-config-macros = { workspace = true }
-ax-cpu = { workspace = true }
-ax-plat = { workspace = true }
-axvisor_api = { workspace = true }
-ax-crate-interface = { workspace = true }
-
-[package.metadata.docs.rs]
-targets = ["riscv64gc-unknown-none-elf"]
