--- platform/riscv64-qemu-virt/linker.lds.S
@@ -1,93 +1,0 @@
-OUTPUT_ARCH(%ARCH%)
-
-BASE_ADDRESS = %KERNEL_BASE%;
-SMP = %SMP%;
-
-ENTRY(_start)
-SECTIONS
-{
-    . = BASE_ADDRESS;
-    _skernel = .;
-
-    .text : ALIGN(4K) {
-        _stext = .;
-        *(.text.boot)
-        *(.text .text.*)
-        . = ALIGN(4K);
-        _etext = .;
-    }
-
-    .rodata : ALIGN(4K) {
-        _srodata = .;
-        *(.rodata .rodata.*)
-        *(.srodata .srodata.*)
-        *(.sdata2 .sdata2.*)
-        . = ALIGN(4K);
-        _erodata = .;
-    }
-
-    .data : ALIGN(4K) {
-        _sdata = .;
-        *(.data.boot_page_table)
-        . = ALIGN(4K);
-        __sdriver_register = .;
-        KEEP(*(.driver.register*))
-        __edriver_register = .;
-
-        *(.data .data.*)
-        *(.sdata .sdata.*)
-        *(.got .got.*)
-    }
-
-    .tdata : ALIGN(0x10) {
-        _stdata = .;
-        *(.tdata .tdata.*)
-        _etdata = .;
-    }
-
-    .tbss : ALIGN(0x10) {
-        _stbss = .;
-        *(.tbss .tbss.*)
-        *(.tcommon)
-        _etbss = .;
-    }
-
-    . = ALIGN(4K);
-    _percpu_start = .;
-    _percpu_end = _percpu_start + SIZEOF(.percpu);
-    .percpu 0x0 : AT(_percpu_start) {
-        _percpu_load_start = .;
-        *(.percpu .percpu.*)
-        _percpu_load_end = .;
-        . = _percpu_load_start + ALIGN(64) * SMP;
-    }
-    . = _percpu_end;
-
-    . = ALIGN(4K);
-    _edata = .;
-
-    .bss : ALIGN(4K) {
-        boot_stack = .;
-        *(.bss.stack)
-        . = ALIGN(4K);
-        boot_stack_top = .;
-
-        _sbss = .;
-        *(.bss .bss.*)
-        *(.sbss .sbss.*)
-        *(COMMON)
-        . = ALIGN(4K);
-        _ebss = .;
-    }
-
-    _ekernel = .;
-
-	/DISCARD/ : {
-        *(.comment) *(.gnu*) *(.note*) *(.eh_frame*)
-    }
-}
-
-SECTIONS {
-    scope_local : { *(scope_local) }
-}
-INSERT AFTER .tbss;
