--- platform/riscv64-qemu-virt/src/lib.rs
@@ -1,41 +1,0 @@
-#![no_std]
-#![cfg(all(target_arch = "riscv64", target_os = "none"))]
-
-#[macro_use]
-extern crate log;
-#[macro_use]
-extern crate ax_plat;
-
-#[cfg(feature = "irq")]
-pub mod irq;
-
-mod boot;
-mod console;
-mod init;
-mod mem;
-mod power;
-mod time;
-
-pub mod config {
-    //! Platform configuration module.
-    //!
-    //! If the `AX_CONFIG_PATH` environment variable is set, it will load the configuration from the specified path.
-    //! Otherwise, it will fall back to the `axconfig.toml` file in the current directory and generate the default configuration.
-    //!
-    //! If the `PACKAGE` field in the configuration does not match the package name, it will panic with an error message.
-    ax_config_macros::include_configs!(path_env = "AX_CONFIG_PATH", fallback = "axconfig.toml");
-    assert_str_eq!(
-        PACKAGE,
-        env!("CARGO_PKG_NAME"),
-        "`PACKAGE` field in the configuration does not match the Package name. Please check your \
-         configuration file."
-    );
-}
-
-pub const fn cpu_count() -> usize {
-    config::plat::MAX_CPU_NUM
-}
-
-pub const fn plic_base() -> usize {
-    config::devices::PLIC_PADDR
-}
