--- platform/riscv64-qemu-virt/src/mem.rs
@@ -1,58 +1,0 @@
-use ax_plat::mem::{MemIf, PhysAddr, RawRange, VirtAddr, pa, va};
-
-use crate::config::{
-    devices::MMIO_RANGES,
-    plat::{KERNEL_BASE_PADDR, PHYS_MEMORY_BASE, PHYS_MEMORY_SIZE, PHYS_VIRT_OFFSET},
-};
-
-struct MemIfImpl;
-
-#[impl_plat_interface]
-impl MemIf for MemIfImpl {
-    /// Returns all physical memory (RAM) ranges on the platform.
-    ///
-    /// All memory ranges except reserved ranges (including the kernel loaded
-    /// range) are free for allocation.
-    fn phys_ram_ranges() -> &'static [RawRange] {
-        // TODO: paser dtb to get the available memory ranges
-        // We can't directly use `PHYS_MEMORY_BASE` here, because it may has been used by sbi.
-        &[(
-            KERNEL_BASE_PADDR,
-            PHYS_MEMORY_BASE + PHYS_MEMORY_SIZE - KERNEL_BASE_PADDR,
-        )]
-    }
-
-    /// Returns all reserved physical memory ranges on the platform.
-    ///
-    /// Reserved memory can be contained in [`phys_ram_ranges`], they are not
-    /// allocatable but should be mapped to kernel's address space.
-    ///
-    /// Note that the ranges returned should not include the range where the
-    /// kernel is loaded.
-    fn reserved_phys_ram_ranges() -> &'static [RawRange] {
-        &[]
-    }
-
-    /// Returns all device memory (MMIO) ranges on the platform.
-    fn mmio_ranges() -> &'static [RawRange] {
-        &MMIO_RANGES
-    }
-
-    /// Translates a physical address to a virtual address.
-    fn phys_to_virt(paddr: PhysAddr) -> VirtAddr {
-        va!(paddr.as_usize() + PHYS_VIRT_OFFSET)
-    }
-
-    /// Translates a virtual address to a physical address.
-    fn virt_to_phys(vaddr: VirtAddr) -> PhysAddr {
-        pa!(vaddr.as_usize() - PHYS_VIRT_OFFSET)
-    }
-
-    /// Returns the kernel address space base virtual address and size.
-    fn kernel_aspace() -> (VirtAddr, usize) {
-        (
-            va!(crate::config::plat::KERNEL_ASPACE_BASE),
-            crate::config::plat::KERNEL_ASPACE_SIZE,
-        )
-    }
-}
